split visualizer into backends; templates directory tree reworked
