+++ jumper_extension/templates/visualizer/__init__.py
@@ -1,0 +1,1 @@
+"""HTML/CSS templates used for JUmPER reports and visual outputs."""
