+++ jumper_extension/templates/visualizer/plotly/add_panel_button/add_panel_button.html
