+++ jumper_extension/templates/visualizer/plotly/cell_range_slider/cell_range_slider.css
