+++ jumper_extension/templates/visualizer/plotly/main.js
