+++ jumper_extension/templates/visualizer/plotly/panel/panel.js
