+++ jumper_extension/templates/visualizer/plotly/show_idle_checkbox/show_idle_checkbox.css
