+++ jumper_extension/templates/visualizer/plotly/styles.css
