+++ jumper_extension/templates/visualizer/plotly/toolbar/toolbar.html
