+++ jumper_extension/templates/visualizer/plotly/visualizer.html
